Moved common files to com.beust.kobalt package
Moved files that were originally under com.beust.kobalt.plugin.java
package to com.beust.kobalt package so that all plugins can access
those classes. Most probably, we should subsequently move these to
com.beust.kobalt.api or com.beust.kobalt.common package.
