+++ src/main/kotlin/com/beust/kobalt/JavaInfo.kt
@@ -1,4 +1,4 @@
-package com.beust.kobalt.plugin.java
+package com.beust.kobalt
 
 import com.beust.kobalt.OperatingSystem
