+++ src/main/kotlin/com/beust/kobalt/Jvm.kt
@@ -1,4 +1,4 @@
-package com.beust.kobalt.plugin.java
+package com.beust.kobalt
 
 import com.beust.kobalt.misc.KobaltLogger
 import com.beust.kobalt.maven.KobaltException
