+++ src/main/kotlin/com/beust/kobalt/internal/JUnitRunner.kt
@@ -4,8 +4,8 @@ import com.beust.kobalt.api.Project
 import com.beust.kobalt.maven.IClasspathDependency
 import com.beust.kobalt.misc.KFiles
 import com.beust.kobalt.misc.KobaltLogger
-import com.beust.kobalt.plugin.java.JavaInfo
-import com.beust.kobalt.plugin.java.SystemProperties
+import com.beust.kobalt.JavaInfo
+import com.beust.kobalt.SystemProperties
 import java.io.File
 
 public class JUnitRunner(override val project: Project, override val classpath: List<IClasspathDependency>)
