+++ src/main/kotlin/com/beust/kobalt/misc/KFiles.kt
@@ -4,7 +4,7 @@ import com.beust.kobalt.api.Kobalt
 import com.beust.kobalt.homeDir
 import com.beust.kobalt.kotlin.BuildFile
 import com.beust.kobalt.maven.KobaltException
-import com.beust.kobalt.plugin.java.SystemProperties
+import com.beust.kobalt.SystemProperties
 import java.io.File
 import java.io.IOException
 import java.io.InputStream
