+++ src/main/kotlin/com/beust/kobalt/plugin/java/JavaPlugin.kt
@@ -11,6 +11,9 @@ import com.beust.kobalt.maven.*
 import com.beust.kobalt.misc.KFiles
 import com.beust.kobalt.misc.KobaltExecutors
 import com.beust.kobalt.misc.KobaltLogger
+import com.beust.kobalt.JavaInfo
+import com.beust.kobalt.Jvm
+import com.beust.kobalt.SystemProperties
 import java.io.File
 import javax.inject.Inject
 import javax.inject.Singleton
